+++ pages/_app.tsx
@@ -7,7 +7,7 @@ import { Toaster } from "react-hot-toast";
 import Navbar from "../components/Navbar";
 import Footer from "../components/Footer";
 import LoadingPage from "../components/LoadingPage";
-import Signature from "../components/Signature";
+import Signature from "../lib/utils/Signature";
 
 const SessionHandler: React.FC<{ children: React.ReactNode }> = ({
   children,
